fix: resolve typing issues and build errors
