+++ app/api/auth/[...nextauth]/route.ts
@@ -1,117 +1,5 @@
-import NextAuth, { type NextAuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import GoogleProvider from "next-auth/providers/google";
-import { MongoDBAdapter } from "@auth/mongodb-adapter";
-import bcrypt from "bcryptjs";
-import clientPromise from "@/lib/mongodb";
-import dbConnect from "@/lib/mongoose";
-import User from "@/models/User";
-
-declare module "next-auth" {
-  interface Session {
-    user: {
-      id: string;
-      name?: string | null;
-      email?: string | null;
-      image?: string | null;
-    }
-  }
-}
-
-export const authOptions: NextAuthOptions = {
-  adapter: MongoDBAdapter(clientPromise),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-      allowDangerousEmailAccountLinking: true,
-    }),
-    CredentialsProvider({
-      name: "credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Mot de passe", type: "password" }
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          throw new Error('Veuillez remplir tous les champs');
-        }
-
-        await dbConnect();
-
-        const user = await User.findOne({ email: credentials.email });
-
-        if (!user) {
-          throw new Error('Aucun utilisateur trouvé avec cet email');
-        }
-
-        if (!user.password) {
-          throw new Error('Veuillez vous connecter avec Google');
-        }
-
-        const isPasswordValid = await bcrypt.compare(
-          credentials.password,
-          user.password
-        );
-
-        if (!isPasswordValid) {
-          throw new Error('Mot de passe incorrect');
-        }
-
-        return {
-          id: user._id.toString(),
-          email: user.email,
-          name: user.name,
-          image: user.image
-        };
-      }
-    })
-  ],
-  callbacks: {
-    async signIn({ user, account, profile }) {
-      if (account?.provider === "google") {
-        try {
-          await dbConnect();
-          const existingUser = await User.findOne({ email: user.email });
-          
-          if (existingUser) {
-            // Mettre à jour les informations Google si nécessaire
-            existingUser.name = user.name;
-            existingUser.image = user.image;
-            await existingUser.save();
-          }
-        } catch (error) {
-          console.error("Erreur lors de la vérification de l'utilisateur:", error);
-          return false;
-        }
-      }
-      return true;
-    },
-    async jwt({ token, user, account }) {
-      if (user) {
-        token.id = user.id;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      return {
-        ...session,
-        user: {
-          ...session.user,
-          id: token.sub
-        }
-      };
-    }
-  },
-  pages: {
-    signIn: '/auth/login',
-    error: '/auth/login'
-  },
-  session: {
-    strategy: "jwt"
-  },
-  secret: process.env.NEXTAUTH_SECRET
-};
+import NextAuth from "next-auth";
+import { authOptions } from "./auth-options";
 
 const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST }; 
+export { handler as GET, handler as POST };
