--- lib/mongodb.ts
@@ -1,28 +1,0 @@
-import { MongoClient } from 'mongodb';
-
-if (!process.env.MONGODB_URI) {
-  throw new Error('Veuillez définir la variable d\'environnement MONGODB_URI');
-}
-
-const uri = process.env.MONGODB_URI;
-let client: MongoClient;
-let clientPromise: Promise<MongoClient>;
-
-if (process.env.NODE_ENV === 'development') {
-  // En développement, utilisez une variable globale pour que la connexion persiste
-  let globalWithMongo = global as typeof globalThis & {
-    _mongoClientPromise?: Promise<MongoClient>;
-  };
-
-  if (!globalWithMongo._mongoClientPromise) {
-    client = new MongoClient(uri);
-    globalWithMongo._mongoClientPromise = client.connect();
-  }
-  clientPromise = globalWithMongo._mongoClientPromise;
-} else {
-  // En production, il est préférable de ne pas utiliser une variable globale
-  client = new MongoClient(uri);
-  clientPromise = client.connect();
-}
-
-export default clientPromise;
